backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
pymongo==4.6.0
chromadb==0.4.15

# Data Validation & Processing
pydantic==2.5.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment & Configuration
python-dotenv==1.0.0

# File Processing
PyPDF2==3.0.1
python-docx==1.2.0

# AI/ML - OpenAI Integration
openai==1.3.0

# Data Analysis
numpy==1.24.3
pandas==2.0.3